% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{exampleMnist}
\alias{exampleMnist}
\title{Example on MNIST dataset
All credit to https://tensorflow.rstudio.com/keras/}
\usage{
exampleMnist()
}
\description{
Example on MNIST dataset
All credit to https://tensorflow.rstudio.com/keras/
}
